alembic/versions/13f8003faa15_encrypt_email_and_student_number.py
"""Encrypt email and student_number

Revision ID: 13f8003faa15
Revises: 
Create Date: 2026-04-03 05:39:44.154114

"""
from typing import Sequence, Union

from alembic import op
import sqlalchemy as sa


# revision identifiers, used by Alembic.
revision: str = '13f8003faa15'
down_revision: Union[str, None] = None
branch_labels: Union[str, Sequence[str], None] = None
depends_on: Union[str, Sequence[str], None] = None


def upgrade() -> None:
    pass


def downgrade() -> None:
    pass
