requirements.txt
Flask==3.1.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Bcrypt==1.0.1
Flask-Limiter==4.1.1
PyMySQL==1.1.1
email-validator==2.2.0
python-dotenv==1.0.1
cryptography==46.0.6
gunicorn==22.0.0
bleach==6.3.0
pyotp==2.9.0
qrcode[pil]==8.2
pytest==8.3.2
pytest-flask==1.3.0
reportlab==4.2.2
Flask-JWT-Extended==4.7.1
Flask-CORS==6.0.2
alembic==1.13.1
